DATAAERO-1 - Further work on an initial prototype of for the template and repositories.
Introduced a dependency to Spring Data Key value to benefit from the infrastructure in place there. Implemented an AerospikeKeyValueAdapter to forward the API interactions triggered by the KeyValueTempalte to the AerospikeClient.

Added rudimentary (but highly incomplete implementations of) BeanDefinitionParsers to start an XML namespace.

Initial draft of a MappingAerospikeConverter to convert domain objects into an array of Bins and read domain objects from records.